backend/requirements.txt
flask
pymongo
passlib[bcrypt]
python-dotenv
python-binance
cryptography
